Move tina-self-hosted-tina config
