+++ examples/tina-self-hosted-demo/graphql.config.js
@@ -1,10 +1,10 @@
 module.exports = {
   projects: {
     app: {
-      schema: ['.tina/__generated__/schema.gql'],
+      schema: ['tina/__generated__/schema.gql'],
       documents: [
-        '.tina/__generated__/queries.gql',
-        '.tina/__generated__/frags.gql',
+        'tina/__generated__/queries.gql',
+        'tina/__generated__/frags.gql',
         '.tina/queries/queries.gql',
         '.tina/queries/frags.gql',
         'pages/**/*.{graphql,js,ts,jsx,tsx}',
