+++ examples/tina-self-hosted-demo/lib/databaseConnection.ts
@@ -1,5 +1,5 @@
-import database from '../.tina/database'
-import { queries } from '../.tina/__generated__/types'
+import database from '../tina/database'
+import { queries } from '../tina/__generated__/types'
 import { resolve } from '@tinacms/datalayer'
 import type { TinaClient } from 'tinacms/dist/client'
