+++ examples/tina-self-hosted-demo/tina/config.tsx
