+++ examples/tina-self-hosted-demo/tina/database.ts
@@ -3,7 +3,8 @@ import { MongodbLevel } from 'mongodb-level'
 import { Octokit } from '@octokit/rest'
 import { Base64 } from 'js-base64'
 
-const isLocal = process.env.TINA_PUBLIC_IS_LOCAL === 'true'
+// const isLocal = process.env.TINA_PUBLIC_IS_LOCAL === 'true'
+const isLocal = true
 
 if (isLocal) console.log('Running TinaCMS in local mode.')
 else console.log('Running TinaCMS in production mode.')
